requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
typing_extensions==4.12.2
Werkzeug==3.1.3
XlsxWriter==3.2.2